requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
importlib-metadata==4.11.3
iniconfig==1.1.1
packaging==21.3
pep8==1.7.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.0
starkbank-ecdsa==2.0.3
tomli==2.0.1
typing_extensions==4.1.1
zipp==3.7.0
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
coverage==6.3.2
freezegun==1.2.1
importlib-metadata==4.11.3
iniconfig==1.1.1
mock==4.0.3
packaging==21.3
pep8==1.7.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.0
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
starkbank-ecdsa==2.0.3
tomli==2.0.1
typing_extensions==4.1.1
zipp==3.7.0
